requirements.txt
msgpack==1.1.0
numpy==2.2.4
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
pyzmq==26.4.0
shiboken6==6.9.0
